Bo sung them thong tin admin o goc tren ben phai mot so trang, them chuyen trang quanLySanPham, mo trang themSanPham

+++ src/main/java/com/example/SellPhone/Model/Product.java
@@ -35,10 +35,10 @@ public class Product {
     private String color;
 
     @Column(name = "import_price")
-    private Double importPrice;
+    private Long importPrice;
 
     @Column(name = "selling_price")
-    private Double sellingPrice;
+    private Long sellingPrice;
 
     @Column(name = "quantity")
     private Integer quantity;
